Update telegram.go for support `message_thread_id` param (#273)
## Summary
I noticed that the current Telegram notifier does not support the
`message_thread_id` parameter, which is necessary for sending messages
to specific threads in a forum-style chat.

## What I did
I added support for the `message_thread_id` parameter to the notifier.

## Note
I'm relatively new to Golang, so I welcome feedback on both
functionality and code style.

---------

Co-authored-by: Jason Lee <huacnlee@gmail.com>